added new room component but not yet complete

+++ LuxuryTravelConcierge/src/main/java/com/revature/models/Booking.java
@@ -11,7 +11,7 @@ public class Booking {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private long bookingId;
+    private int bookingId;
 
     //Will use a foreign key to room id later
     private int roomId;
@@ -49,7 +49,7 @@ public class Booking {
         return bookingId;
     }
 
-    public void setBookingId(long bookingId) {
+    public void setBookingId(int bookingId) {
         this.bookingId = bookingId;
     }
